chore(suite): move GitHub release notes helpers from suite package
