+++ packages/urls/src/index.ts
@@ -1,1 +1,2 @@
 export * from './urls';
+export * from './github';
